Fixed EventHandlers Fixed Multiple Connection Issues Fixed Code

+++ FiroozehGameService/Models/GSLive/RT/Packet.cs
@@ -34,7 +34,7 @@ namespace FiroozehGameService.Models.GSLive.RT
         internal int Action;
         internal long ClientReceiveTime;
         internal long ClientSendTime;
-        internal ulong Hash;
+        internal long Hash;
         internal byte[] Payload;
         internal GProtocolSendType SendType;
 
@@ -43,7 +43,7 @@ namespace FiroozehGameService.Models.GSLive.RT
             Deserialize(buffer);
         }
 
-        public Packet(ulong hash, int action, GProtocolSendType sendType = GProtocolSendType.UnReliable,
+        public Packet(long hash, int action, GProtocolSendType sendType = GProtocolSendType.UnReliable,
             byte[] payload = null)
         {
             Hash = hash;
@@ -83,7 +83,7 @@ namespace FiroozehGameService.Models.GSLive.RT
 
                 // data Segment
                 packetWriter.Write((byte) SendType);
-                packetWriter.Write(Hash);
+                packetWriter.Write((ulong) Hash);
 
                 if (havePayload == 0x1) packetWriter.Write(Payload);
                 if (haveSendTime == 0x1) packetWriter.Write(ClientSendTime);
@@ -104,7 +104,7 @@ namespace FiroozehGameService.Models.GSLive.RT
                 if (havePayload == 0x1) _payloadLen = packetWriter.ReadUInt16();
 
                 SendType = (GProtocolSendType) packetWriter.ReadByte();
-                Hash = packetWriter.ReadUInt64();
+                Hash = (long) packetWriter.ReadUInt64();
 
                 if (havePayload == 0x1) Payload = packetWriter.ReadBytes(_payloadLen);
                 if (haveSendTime == 0x1) ClientSendTime = packetWriter.ReadInt64();
